Move event dispatcher into a decorator
