+++ src/GuzzleHttpAdapter.php
@@ -53,7 +53,7 @@ class GuzzleHttpAdapter extends AbstractCurlHttpAdapter
     /**
      * {@inheritdoc}
      */
-    protected function doSendInternalRequest(InternalRequestInterface $internalRequest)
+    protected function sendInternalRequest(InternalRequestInterface $internalRequest)
     {
         try {
             $response = $this->createRequest($internalRequest)->send();
@@ -81,7 +81,7 @@ class GuzzleHttpAdapter extends AbstractCurlHttpAdapter
     /**
      * {@inheritdoc}
      */
-    protected function doSendInternalRequests(array $internalRequests, $success, $error)
+    protected function sendInternalRequests(array $internalRequests, $success, $error)
     {
         $requests = array();
         foreach ($internalRequests as $internalRequest) {
