+++ src/StopwatchHttpAdapter.php
@@ -11,7 +11,6 @@
 
 namespace Ivory\HttpAdapter;
 
-use Psr\Http\Message\RequestInterface;
 use Symfony\Component\Stopwatch\Stopwatch;
 
 /**
@@ -19,132 +18,33 @@ use Symfony\Component\Stopwatch\Stopwatch;
  *
  * @author GeLo <geloen.eric@gmail.com>
  */
-class StopwatchHttpAdapter extends AbstractHttpAdapterTemplate
+class StopwatchHttpAdapter extends PsrHttpAdapterDecorator
 {
-    /** @var \Ivory\HttpAdapter\HttpAdapterInterface */
-    private $httpAdapter;
-
     /** @var \Symfony\Component\Stopwatch\Stopwatch */
     private $stopwatch;
 
     /**
-     * Creates a stopwatch event subscriber.
+     * Creates a stopwatch http adapter.
      *
      * @param \Ivory\HttpAdapter\HttpAdapterInterface $httpAdapter The http adapter.
      * @param \Symfony\Component\Stopwatch\Stopwatch  $stopwatch   The stopwatch.
      */
     public function __construct(HttpAdapterInterface $httpAdapter, Stopwatch $stopwatch)
     {
-        $this->setHttpAdapter($httpAdapter);
-        $this->setStopwatch($stopwatch);
-    }
+        parent::__construct($httpAdapter);
 
-    /**
-     * Gets the http adapter.
-     *
-     * @return \Ivory\HttpAdapter\HttpAdapterInterface The http adapter.
-     */
-    public function getHttpAdapter()
-    {
-        return $this->httpAdapter;
-    }
-
-    /**
-     * Sets the http adapter.
-     *
-     * @param \Ivory\HttpAdapter\HttpAdapterInterface $httpAdapter The http adapter.
-     */
-    public function setHttpAdapter(HttpAdapterInterface $httpAdapter)
-    {
-        $this->httpAdapter = $httpAdapter;
-    }
-
-    /**
-     * Gets the stopwatch.
-     *
-     * @return \Symfony\Component\Stopwatch\Stopwatch The stopwatch.
-     */
-    public function getStopwatch()
-    {
-        return $this->stopwatch;
-    }
-
-    /**
-     * Sets the stopwatch.
-     *
-     * @param \Symfony\Component\Stopwatch\Stopwatch $stopwatch The stopwatch.
-     */
-    public function setStopwatch(Stopwatch $stopwatch)
-    {
         $this->stopwatch = $stopwatch;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getConfiguration()
+    protected function decorate($method, array $params = array())
     {
-        return $this->httpAdapter->getConfiguration();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setConfiguration(ConfigurationInterface $configuration)
-    {
-        $this->httpAdapter->setConfiguration($configuration);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function send($uri, $method, array $headers = array(), $datas = array(), array $files = array())
-    {
-        return $this->watch('send', array($uri, $method, $headers, $datas, $files));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function sendRequest(RequestInterface $request)
-    {
-        return $this->watch('sendRequest', array($request));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function sendRequests(array $requests)
-    {
-        return $this->watch('sendRequests', array($requests));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return $this->httpAdapter->getName();
-    }
-
-    /**
-     * Watches a method.
-     *
-     * @param string $method The method.
-     * @param array  $params The parameters.
-     *
-     * @throws \Exception If an exception has been thrown during the HTTP adapter call.
-     *
-     * @return mixed The result.
-     */
-    private function watch($method, array $params = array())
-    {
-        $name = 'ivory.http_adapter';
-
-        $this->stopwatch->start($name);
+        $this->stopwatch->start($name = 'ivory.http_adapter');
 
         try {
-            $result = call_user_func_array(array($this->httpAdapter, $method), $params);
+            $result = parent::decorate($method, $params);
         } catch (\Exception $e) {
             $this->stopwatch->stop($name);
