+++ src/Zend1HttpAdapter.php
@@ -48,7 +48,7 @@ class Zend1HttpAdapter extends AbstractHttpAdapter
     /**
      * {@inheritdoc}
      */
-    protected function doSendInternalRequest(InternalRequestInterface $internalRequest)
+    protected function sendInternalRequest(InternalRequestInterface $internalRequest)
     {
         $this->client
             ->resetParameters(true)
