+++ src/Zend2HttpAdapter.php
@@ -50,7 +50,7 @@ class Zend2HttpAdapter extends AbstractHttpAdapter
     /**
      * {@inheritdoc}
      */
-    protected function doSendInternalRequest(InternalRequestInterface $internalRequest)
+    protected function sendInternalRequest(InternalRequestInterface $internalRequest)
     {
         $this->client
             ->resetParameters(true)
